Move lifecycle calls to base fragment
